+++ app-mobile/src/main/java/ru/radiationx/anilibria/presentation/release/details/ReleaseInfoPresenter.kt
@@ -1,24 +1,23 @@
 package ru.radiationx.anilibria.presentation.release.details
 
 import android.util.Log
-import com.arellomobile.mvp.InjectViewState
-import ru.radiationx.anilibria.entity.app.release.ReleaseFull
-import ru.radiationx.anilibria.entity.app.release.ReleaseItem
-import ru.radiationx.anilibria.entity.app.vital.VitalItem
-import ru.radiationx.anilibria.entity.common.AuthState
-import ru.radiationx.anilibria.model.data.remote.Api
-import ru.radiationx.anilibria.model.data.remote.address.ApiConfig
-import ru.radiationx.anilibria.model.data.remote.api.PageApi
-import ru.radiationx.anilibria.model.interactors.ReleaseInteractor
-import ru.radiationx.anilibria.model.repository.AuthRepository
-import ru.radiationx.anilibria.model.repository.FavoriteRepository
-import ru.radiationx.anilibria.model.repository.HistoryRepository
-import ru.radiationx.anilibria.model.repository.VitalRepository
+import moxy.InjectViewState
 import ru.radiationx.anilibria.navigation.Screens
 import ru.radiationx.anilibria.presentation.common.BasePresenter
 import ru.radiationx.anilibria.presentation.common.IErrorHandler
 import ru.radiationx.anilibria.presentation.common.ILinkHandler
 import ru.radiationx.anilibria.utils.Utils
+import ru.radiationx.data.datasource.remote.address.ApiConfig
+import ru.radiationx.data.datasource.remote.api.PageApi
+import ru.radiationx.data.entity.app.release.ReleaseFull
+import ru.radiationx.data.entity.app.release.ReleaseItem
+import ru.radiationx.data.entity.app.vital.VitalItem
+import ru.radiationx.data.entity.common.AuthState
+import ru.radiationx.data.interactors.ReleaseInteractor
+import ru.radiationx.data.repository.AuthRepository
+import ru.radiationx.data.repository.FavoriteRepository
+import ru.radiationx.data.repository.HistoryRepository
+import ru.radiationx.data.repository.VitalRepository
 import ru.terrakok.cicerone.Router
 import javax.inject.Inject
 
@@ -139,7 +138,7 @@ class ReleaseInfoPresenter @Inject constructor(
     fun onClickWatchWeb() {
         currentData?.let { release ->
             release.moonwalkLink?.let {
-                viewState.playWeb(it)
+                viewState.playWeb(it, release.code.orEmpty())
             }
         }
     }
@@ -179,7 +178,7 @@ class ReleaseInfoPresenter @Inject constructor(
 
     fun onClickDonate() {
         //router.navigateTo(Screens.StaticPage(PageApi.PAGE_ID_DONATE))
-        Utils.externalLink("${apiConfig.baseUrl}/${PageApi.PAGE_ID_DONATE}")
+        Utils.externalLink("${apiConfig.siteUrl}/${PageApi.PAGE_PATH_DONATE}")
     }
 
     fun onClickFav() {
@@ -243,7 +242,7 @@ class ReleaseInfoPresenter @Inject constructor(
 
     fun onDialogDonateClick() {
         //router.navigateTo(Screens.StaticPage(PageApi.PAGE_ID_DONATE))
-        Utils.externalLink("${apiConfig.baseUrl}/${PageApi.PAGE_ID_DONATE}")
+        Utils.externalLink("${apiConfig.siteUrl}/${PageApi.PAGE_PATH_DONATE}")
     }
 
     fun onResetEpisodesHistoryClick() {
